// src/lib/api/generated/model/technicalDeliveryPageSchema.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The canonical backend API for COMPASS clients.
 * OpenAPI spec version: 1.0.0
 */
import type { TechnicalDeliverySchema } from './technicalDeliverySchema';

export interface TechnicalDeliveryPageSchema {
  items: TechnicalDeliverySchema[];
  page: number;
  page_size: number;
  total: number;
}